Moved out of the __init__.py the functions: tooltipWithColour()
